+++ build.zig
@@ -298,6 +298,10 @@ fn addWasmRenderer(
         "errorCode",
         "releaseSource",
         "releaseOutput",
+        "allocateMacroConfig",
+        "configureMathMacros",
+        "releaseMacroConfig",
+        "clearMathMacros",
     };
     const renderer = b.addExecutable(.{
         .name = name,